+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Build 1522 - 2026-09-25
+
+- automated build
+
 ## Build 1521 - 2026-09-25
 
 - automated build